backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.12
pydantic==2.9.2
pydantic-settings==2.5.2
supabase==2.9.0
openai==1.51.0
google-generativeai==0.8.3
pytesseract==0.3.13
Pillow>=11.0.0
pdfplumber==0.11.4
pdf2image==1.17.0
python-dotenv==1.0.1
httpx==0.27.2
